WIP: Cleanup and implement proper timing.

+++ c++/DrumSequencer/miditime.h
@@ -25,7 +25,7 @@ namespace MidiTime
 {
 DWORD BPMToMicroseconds(DWORD bpm);
 DWORD MicrosecondsToBPM(DWORD microseconds);
-BYTE DurationToTicks(Duration duration);
+DWORD DurationToTicks(Duration duration);
 Duration TicksToDuration(BYTE Ticks);
 };
